signatures: add Maxine Aubrey
Nix is on the brink due to the apathetic stance of the foundation towards many
community problems, and the continued pursuit of neutrality and centrist balance
on basic humanity. I would much rather see this result in the community finally
being able to address this on-going situation than it cause a fork, but any path
forward is preferable to the current status quo.