Update signatures.md
I too sign this, albeit reluctantly.

To any potential employers or HR looking at this... I've been in the loop, and there has been an issue with the founder for well over a decade. Everyone that is in the loop knows. I just want to clarify that this is not in bad faith. This is in the best interests of an open source project.